Add in restrictions for only beta testers

+++ app/Http/Controllers/UKCP/Token.php
@@ -22,6 +22,8 @@ class Token extends BaseController
 
     public function create()
     {
+        $this->authorize('ukcp-beta');
+
         $currentTokens = $this->ukcp->getValidTokensFor(auth()->user());
 
         if ($currentTokens->count() >= 4) {
@@ -45,6 +47,8 @@ class Token extends BaseController
 
     public function show()
     {
+        $this->authorize('ukcp-beta');
+
         $latestId = $this->ukcp->getValidTokensFor(auth()->user());
 
         if ($latestId->isEmpty()) {
@@ -56,6 +60,8 @@ class Token extends BaseController
 
     public function destroy($tokenId)
     {
+        $this->authorize('ukcp-beta');
+
         $delete = $this->ukcp->deleteToken($tokenId);
 
         if (!$delete) {
@@ -69,6 +75,8 @@ class Token extends BaseController
 
     public function download($tokenId)
     {
+        $this->authorize('ukcp-beta');
+
         $tokenPath = storage_path('app/ukcp/tokens/') . auth()->user()->id . '/' . $tokenId . '.json';
         $headers = array(
             'Content-Type: application/json',
